introduce util for indent root cell mappers
