+++ cell/src/main/java/jetbrains/jetpad/cell/toView/IndentRootCellMapper.java
@@ -16,7 +16,6 @@
 package jetbrains.jetpad.cell.toView;
 
 import com.google.common.base.Strings;
-import jetbrains.jetpad.base.Handler;
 import jetbrains.jetpad.base.Registration;
 import jetbrains.jetpad.cell.Cell;
 import jetbrains.jetpad.cell.CellPropertySpec;
@@ -28,6 +27,7 @@ import jetbrains.jetpad.cell.indent.updater.IndentUpdaterTarget;
 import jetbrains.jetpad.cell.mappers.CellMapper;
 import jetbrains.jetpad.cell.toUtil.AncestorUtil;
 import jetbrains.jetpad.cell.toUtil.CounterUtil;
+import jetbrains.jetpad.cell.indent.IndentUtil;
 import jetbrains.jetpad.mapper.MappingContext;
 import jetbrains.jetpad.model.collections.CollectionItemEvent;
 import jetbrains.jetpad.model.collections.set.ObservableSet;
@@ -149,42 +149,13 @@ class IndentRootCellMapper extends BaseCellMapper<IndentCell, VerticalView> {
       @Override
       public void propertyChanged(Cell cell, final CellPropertySpec<?> prop, final PropertyChangeEvent<?> event) {
         if (CounterUtil.isCounterProp(prop)) {
-          iterateLeaves(cell, new Handler<Cell>() {
-            @Override
-            public void handle(Cell item) {
-              BaseCellMapper<?, ?> mapper = (BaseCellMapper<?, ?>) getDescendantMapper(item);
-              if (mapper == null) {
-                throw new IllegalStateException();
-              }
-              if (CounterUtil.update(mapper, prop, event)) {
-                mapper.refreshProperties();
-              }
-            }
-          });
+          IndentUtil.updateCounters(cell, prop, event, IndentRootCellMapper.this);
         } else if (prop == Cell.VISIBLE) {
           myIndentUpdater.visibilityChanged(cell, (PropertyChangeEvent<Boolean>) event);
         } else if (Cell.isPopupProp(prop)) {
           updateIndentCellPopup(cell, (PropertyChangeEvent<Cell>) event);
         } else if (prop == Cell.BACKGROUND) {
-          iterateLeaves(cell, new Handler<Cell>() {
-            @Override
-            public void handle(Cell item) {
-              BaseCellMapper itemMapper = (BaseCellMapper) getDescendantMapper(item);
-              itemMapper.setAncestorBackground(AncestorUtil.getAncestorBackground(getSource(), item));
-              itemMapper.refreshProperties();
-            }
-          });
-        }
-      }
-
-      private void iterateLeaves(Cell cell, Handler<Cell> handler) {
-        for (Cell child : cell.children()) {
-          if (!Composites.isVisible(child)) continue;
-          if (child instanceof IndentCell) {
-            iterateLeaves(child, handler);
-          } else {
-            handler.handle(child);
-          }
+          IndentUtil.updateBackground(cell, IndentRootCellMapper.this);
         }
       }
     });
